move scenewidgets to their own class
