+++ meerk40t/gui/scenewidgets/rectselectwidget.py
