+++ meerk40t/gui/scenewidgets/reticlewidget.py
@@ -1,4 +1,3 @@
-
 import wx
 
 from meerk40t.gui.laserrender import DRAW_MODE_RETICLE
